% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglor.R
\name{gglor}
\alias{gglor}
\title{Plot log odds ratio}
\usage{
gglor(vol3D, cond, corr, yint, ychar, xchar, xnumbreaks, xlabbreaks)
}
\arguments{
\item{vol3D}{Input 3D volume.}

\item{cond}{Condition of the lag-two volume. Options: marginal (.), positive (1), or negative (0).}

\item{corr}{Continuity correction for the odds ratio.}

\item{yint}{y-axis limits.}

\item{ychar}{y-axis label.}

\item{xchar}{x-axis label.}

\item{xnumbreaks}{x-axis breaks.}

\item{xlabbreaks}{x-axis label breaks.}
}
\value{
A ggplot2 object depicting the log odds ratio over the first dimension.
}
\description{
This function generates a ggplot graph with a log odds ratio
between a 3D volumes and its lag volume and conditioned by its lag-two. 
Computation is over the fist dimension (time) the x-axis.
}
\details{
The log odds ratio is computed over the first dimension of the input 3D volume.
}
\examples{

# load("data/Y365.rda")
# gglor(vol3D = Y365,
#      cond = ".",
#      corr = 0.5,
#      yint = c(1.1, 3.6),
#      ychar = expression(LOR[t]),
#      xchar = "t (year)",
#      xnumbreaks = c(1, 21, 41, 61),
#      xlabbreaks = c("1 (1960)", "21 (1980)", "41 (2000)", "61 (2020)"))

}
\author{
Zeus Gracia-Tabuenca
}
