% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I.weak.record.R
\name{I.weak.record}
\alias{I.weak.record}
\title{Record Indicators}
\usage{
I.weak.record(X)
}
\arguments{
\item{X}{The vector}
}
\value{
The binary vector of the same length as \code{X}, 
  indicating the record occurrence.
}
\description{
This function returns the record indicators of the values in a 
  vector. The record indicator for each value in a vector is a binary 
  variable that takes the value 1 if the corresponding value in the vector 
  is a record and 0 otherwise. Indicators for \eqn{r}-tied records have a 
  value \eqn{1 / r}.
}
\author{
Jorge Castillo-Mateo
}
