% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Metrics for Validation}
\usage{
metrics(
  Y,
  p,
  year,
  day,
  site,
  metric = c("point", "N", "DIC"),
  D1 = 2:(62/2),
  D2 = (62/2 + 1):62,
  B = 2000
)
}
\arguments{
\item{Y}{The observed values (vector)}

\item{p}{The estimated probabilities (vector or matrix with rows being 
replicates)}

\item{year, day, site}{The year, day, and site associated with the values 
(vectors) (\code{years} must be in order from lower to higher)}

\item{metric}{A character string \code{"point"}, \code{"N"}, or \code{"DIC"}}

\item{D1, D2}{Two shorter year periods on which to calculate the metrics, in 
addition to the period of all years}

\item{B}{If \code{p} is a vector, the number of replicates to calculate AD's}
}
\value{
If \code{metric = "point"}; the metrics (vector):
  \item{BS}{Brier Score}
  \item{BS1}{BS in \code{D1}}
  \item{BS2}{BS in \code{D2}}
  \item{AUC}{Area under the ROC curve}
  \item{AUC1}{AUC in \code{D1}}
  \item{AUC2}{AUC in \code{D2}}
  
  If \code{metric = "N"}; a vector of AD's averaged grouped by year
  
  If \code{metric = "DIC"}; a vector with DIC, D (deviance), and penalty (pD)
}
\description{
This function is used to validate the predicted probabilities and number
  of records with the actual values.
}
\details{
Values different from 0 or 1 (tied records) are removed from the metrics.
  
  The DIC uses mean values instead of removing or sampling values different 
  from 0 or 1.
}
\author{
Jorge Castillo-Mateo
}
