% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rom}
\alias{predict.rom}
\title{Predict Method for Record Occurrence Model Fits}
\usage{
\method{predict}{rom}(
  object,
  newdata,
  newcoords,
  newsite,
  newyear,
  newday,
  type = c("KFCV", "sample")
)
}
\arguments{
\item{object}{Object of class inheriting from "rom"}

\item{newdata}{An optional list with \code{x} in which to look for variables 
with which to predict, \code{coords} and \code{site} in which to predict, 
and \code{day} or \code{year}. 
If omitted, the fitted values are used.}

\item{newcoords}{New coordinates for the sites in \code{newdata$site} or 
\code{newsite}.}

\item{newsite, newyear, newday}{A vector where for each observation includes 
the site, year or day.}

\item{type}{the type of prediction required. For \eqn{K}-fold 
cross-validation \code{"KFCV"} uses observed indicators from \code{data}
and samples weak records from their prior probability.}
}
\value{
A \code{"rom"} list with elements:
  \item{params}{Matrix where rows are MCMC simulations and cols are 
    parameters:
  \deqn{\beta_1,\ldots,\beta_k}}
  \item{\code{y}}{Data fitted}
  \item{\code{x}}{Covariates}
}
\description{
Predicted values based on rom (record occurrence model) object.
}
\details{
Predicted values are on the scale of the response (probabilities).
}
\author{
Jorge Castillo-Mateo
}
