% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rom.R
\name{rom}
\alias{rom}
\title{Record Occurrence Model Fitting}
\usage{
rom(
  formula,
  data,
  coords,
  site,
  year,
  day,
  model = c("1", "2", "3", "4", "5"),
  scale.cov = TRUE,
  decay.prior = c("gamma", "unif"),
  inits = NULL,
  prior = list(beta = 1e-04, sigma = c(2, 1), decay = c(2, 1)),
  n.sims = 1000,
  n.thin = 1,
  n.burnin = 1000,
  n.report = 100
)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the 
model to be fitted. (Has some requirements with \code{trend}, \code{lag1}, 
and \code{lag2}).}

\item{data}{a data frame (\eqn{N \times p}) containing the variables in the 
model.}

\item{coords}{a matrix or data frame (\eqn{n \times 2}) containing the 
coordinates of the sites.}

\item{site, year, day}{vectors (\eqn{N \times 1}) containing the site, year, 
and day the measurements were taken.}

\item{model}{a character indicating the model to fit from the paper.}

\item{scale.cov}{a boolean indicating whether or not the columns of the 
design matrix must be scaled to have zero mean and unit variance.}

\item{decay.prior}{a character string indicating the prior for the decay 
parameter, gamma or uniform in an interval.}

\item{inits}{a vector of initial values (if NULL, random values).}

\item{prior}{a list containing the parameters of the prior distributions.}

\item{n.sims, n.thin, n.burnin, n.report}{(i) Number of iterations not discarded. 
(ii) Thinning rate. 
(iii) Number of iterations discarded at the beginning. 
(iv) Report the number of iterations rate.}
}
\value{
A list with many elements. Samples from the model parameters are in 
  \code{params} where rows are MCMC simulations and cols are parameters.
}
\description{
This function fits the Record Occurrence Model in a Bayesian framework 
  using MCMC.
}
\details{
The fitting algorithm is a data-augmentation Metropolis-within-Gibbs MCMC 
  algorithm. 
  
  \code{formula} requires two variables, \code{lag1} and \code{lag2}, 
  that are the record indicator one and two days before. 
  
  \code{data} requires column names \code{trend} (annual trend), 
  \code{lag1}, and \code{lag2}.
  
  The response variable must be \code{Y}. It can be 0's and 1's indicating 
  the occurrence of a record, but values \eqn{1/r} (\eqn{r=1,,2\ldots}) are 
  allowed indicating the probability of a weak record being a strong record.
  (If all values are 0's and 1's the function could return an error.)
  
  The models for the first and second days of the year are 
  \code{Y ~ trend + lag1}.
}
\note{
It is necessary to have data ordered as using if c() over a 
  3-dimensional array where the first dimension is year, the second is day, 
  and the third is site. This is, first all data for the first site, then 
  for the first day and it progresses by years. (This will also be necessary 
  for prediction.)
}
\author{
Jorge Castillo-Mateo
}
